부산대 BE_박수빈 6주차 과제 (1단계) (#117)
* step0 시작

* refactor : controller/api 수정

* refactor : controller/web 수정

* refactor : DTO 수정

* refactor : model 수정

* refactor : repository 수정

* refactor : service 수정

* refactor : config 수정

* refactor : js 수정

* refactor : html 수정

* docs : readme 수정

+++ src/main/java/gift/dto/CategoryDTO.java
@@ -10,7 +10,7 @@ public class CategoryDTO {
 
     @Schema(description = "카테고리 ID", example = "1")
     @NotBlank(message = "아이디값은 필수 값입니다.")
-    private int id;
+    private Long id;
 
     @Schema(description = "카테고리 이름", example = "교환권")
     @NotBlank(message = "이름은 필수 값입니다.")
@@ -28,7 +28,7 @@ public class CategoryDTO {
     @Size(max = 500, message = "설명은 최대 500자까지 가능합니다.")
     private String description;
 
-    public CategoryDTO(int id, String name, String color, String imageUrl, String description) {
+    public CategoryDTO(Long id, String name, String color, String imageUrl, String description) {
         this.id = id;
         this.name = name;
         this.color = color;
@@ -40,7 +40,7 @@ public class CategoryDTO {
         return new CategoryDTO(category.getId(), category.getName(), category.getColor(), category.getImageUrl(), category.getDescription());
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
@@ -60,7 +60,7 @@ public class CategoryDTO {
         return description;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
